refactor #milestone_response Move from `ApplicationController` to `MilestoneHelper` for reuse.
